feat: access token issue logic change header to cookie
