+++ server/server/src/main/java/com/cactusvilleage/server/auth/repository/OAuth2AuthorizationRequestRepository.java
@@ -29,10 +29,10 @@ public class OAuth2AuthorizationRequestRepository implements AuthorizationReques
             CookieUtil.deleteCookie(request, response, REFRESH_TOKEN);
             return;
         }
-        CookieUtil.addCookie(response, OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME, CookieUtil.serialize(authorizationRequest), cookieExpiresSecond);
+        CookieUtil.addCookie(response, OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME, CookieUtil.serialize(authorizationRequest), cookieExpiresSecond, true);
         String redirectUriAfterLogin = request.getParameter(REDIRECT_URI_PARAM_COOKIE_NAME);
         if (StringUtils.isNotBlank(redirectUriAfterLogin)) {
-            CookieUtil.addCookie(response, REDIRECT_URI_PARAM_COOKIE_NAME, redirectUriAfterLogin, cookieExpiresSecond);
+            CookieUtil.addCookie(response, REDIRECT_URI_PARAM_COOKIE_NAME, redirectUriAfterLogin, cookieExpiresSecond, true);
         }
     }
