+++ server/server/src/main/java/com/cactusvilleage/server/auth/web/oauth/OAuth2AuthenticationSuccessHandler.java
@@ -2,9 +2,7 @@ package com.cactusvilleage.server.auth.web.oauth;
 
 import com.cactusvilleage.server.auth.repository.OAuth2AuthorizationRequestRepository;
 import com.cactusvilleage.server.auth.util.CookieUtil;
-import com.cactusvilleage.server.auth.util.HeaderUtil;
 import com.cactusvilleage.server.global.exception.BusinessLogicException;
-import com.cactusvilleage.server.global.exception.ExceptionCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
@@ -26,14 +24,14 @@ import static com.cactusvilleage.server.global.exception.ExceptionCode.*;
 @RequiredArgsConstructor
 public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
     private final AppProperties appProperties;
-    private final HeaderUtil headerUtil;
+    private final CookieUtil cookieUtil;
     private final OAuth2AuthorizationRequestRepository authorizationRequestRepository;
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
         String targetUri = determineTargetUri(request);
         clearAuthenticationAttributes(request, response);
-        headerUtil.generateTokens(request, response, authentication);
+        cookieUtil.generateTokens(request, response, authentication);
         getRedirectStrategy().sendRedirect(request, response, targetUri);
     }
 
@@ -56,8 +54,6 @@ public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationS
                 .stream()
                 .anyMatch(authorizedRedirectUri -> {
                     URI authorizedURI = URI.create(authorizedRedirectUri);
-//                    return authorizedURI.getHost().equalsIgnoreCase(clientRedirectUri.getHost())
-//                            && authorizedURI.getPort() == clientRedirectUri.getPort();
                     return authorizedURI.getHost().equalsIgnoreCase(clientRedirectUri.getHost());
                 });
     }
